publisher and mergers added
