+++ src/topic/mod.rs
@@ -1,2 +1,3 @@
 pub mod topic;
 pub mod bootstrap;
+pub mod publisher;
